refactor(sparse_matrix): improve implementation
- Unify sparse matrix iterator implementations using templated IsConst parameter
- Add bitset operations for improved bit manipulation efficiency
- Extract common functionality into private helper methods
- Optimize row value counting using std::accumulate
- Improve code organization and readability with smaller, focused methods
- Add validation of matrix indices in a centralized method
- Enhance iterator implementation with cleaner const/non-const handling

This change reduces code duplication and improves maintainability while keeping
the same functionality. Performance is improved through better bit manipulation
and optimized algorithms for counting values.

+++ src/qmutils/sparse_matrix.h
@@ -40,7 +40,10 @@ class SparseRow {
 
   size_t byte_index(size_t pos) const { return pos / BITS_PER_BYTE; }
   size_t bit_index(size_t pos) const { return pos % BITS_PER_BYTE; }
-  uint8_t bit_mask(size_t bit_pos) const { return 1U << bit_pos; }
+
+  uint8_t bit_mask(size_t bit_pos) const {
+    return static_cast<uint8_t>(1U << bit_pos);
+  }
 
  public:
   explicit SparseRow(size_t size = 0) { resize(size); }
